many many changes

+++ tanktrouble/algo/tianshou-train.py
@@ -10,8 +10,11 @@ git+https://github.com/thu-ml/tianshou
 """
 import argparse
 import copy
+import datetime
 import os
 import sys
+from datetime import time
+from math import prod
 from time import sleep
 from typing import Optional, Tuple
 
@@ -20,19 +23,30 @@ import numpy as np
 import pygame
 import torch
 from pettingzoo.utils import parallel_to_aec
-from tianshou.data import Collector, VectorReplayBuffer
-from tianshou.env import DummyVectorEnv
+from tianshou.data import Collector, VectorReplayBuffer, PrioritizedVectorReplayBuffer
+from tianshou.env import DummyVectorEnv, SubprocVectorEnv
 from tianshou.env.pettingzoo_env import PettingZooEnv
-from tianshou.policy import BasePolicy, DQNPolicy, MultiAgentPolicyManager, RandomPolicy, RainbowPolicy
+from tianshou.policy import BasePolicy, DQNPolicy, MultiAgentPolicyManager, RandomPolicy, RainbowPolicy, ICMPolicy
 from tianshou.trainer import OffpolicyTrainer
 from tianshou.utils.net.common import Net
+from tianshou.utils.net.discrete import IntrinsicCuriosityModule
+
 import tanktrouble.env.tanktrouble_env as tanktrouble
+from tanktrouble.models.tt_network import DQN_TT
 
 from league import LeaguePolicy
 
 from pettingzoo.classic import tictactoe_v3
 
 is_distrib = True
+icm = False
+import multiprocessing
+
+s_x = 8
+s_y = 5
+img_size = (s_x * 3, s_y * 3, 5)
+
+watch = False
 
 
 def _get_agents(
@@ -47,34 +61,48 @@ def _get_agents(
         if isinstance(env.observation_space, gymnasium.spaces.Dict)
         else env.observation_space
     )
-    hidden_sizes = [1024, 1024, 1024]
+    hidden_sizes = [512]
     if agent_learn is None:
         # model
-        net = Net(
-            state_shape=gymnasium.spaces.utils.flatdim(observation_space),
-            action_shape=gymnasium.spaces.utils.flatdim(env.action_space),
-            hidden_sizes=hidden_sizes,
-            device="cuda" if torch.cuda.is_available() else "cpu",
-            num_atoms=51 if is_distrib else 1,
-        ).to("cuda" if torch.cuda.is_available() else "cpu")
-        net_fixed = Net(
-            state_shape=gymnasium.spaces.utils.flatdim(observation_space),
-            action_shape=gymnasium.spaces.utils.flatdim(env.action_space),
-            hidden_sizes=hidden_sizes,
-            device="cuda" if torch.cuda.is_available() else "cpu",
-            num_atoms=51 if is_distrib else 1,
-        ).to("cuda" if torch.cuda.is_available() else "cpu")
+        action_dim = gymnasium.spaces.utils.flatdim(env.action_space)
 
-        if optim is None:
-            optim = torch.optim.Adam(net.parameters(), lr=1e-4)
-            optim_fixed = torch.optim.SGD(net_fixed.parameters(), lr=0.0)
+        def make_net(grad=True, features_only=False):
+            # return Net(
+            #     state_shape=gymnasium.spaces.utils.flatdim(observation_space),
+            #     action_shape=gymnasium.spaces.utils.flatdim(env.action_space),
+            #     hidden_sizes=hidden_sizes,
+            #     device="cuda" if torch.cuda.is_available() else "cpu",
+            #     num_atoms=51 if is_distrib else 1,
+            # ).to("cuda" if torch.cuda.is_available() else "cpu")
+            return DQN_TT(
+                rest=gymnasium.spaces.utils.flatdim(observation_space) - (
+                    prod(img_size) if tanktrouble.image_in_obs else 0),
+                action_shape=[action_dim],
+                hidden=hidden_sizes,
+                c=img_size[2] if tanktrouble.image_in_obs else 0,
+                h=img_size[1] if tanktrouble.image_in_obs else 0,
+                w=img_size[0] if tanktrouble.image_in_obs else 0,
+                device="cuda" if torch.cuda.is_available() else "cpu",
+                use_img=tanktrouble.image_in_obs,
+                grad=grad,
+                features_only=features_only,
+                atoms=51 if is_distrib else 1,
+            ).to("cuda" if torch.cuda.is_available() else "cpu")
+
+        net = make_net()
+        net_fixed = make_net(grad=False).train(False)
+
+        optim = torch.optim.Adam(net.parameters(), lr=3e-4)
+        optim_fixed = torch.optim.SGD(net_fixed.parameters(), lr=0.0)
+        icm_optim = torch.optim.Adam(net.parameters(), lr=3e-4)
+
         if not is_distrib:
             agent_learn = DQNPolicy(
                 model=net,
                 optim=optim,
                 # discount_factor=0.98,
-                estimation_step=15,
-                target_update_freq=5000,
+                estimation_step=5,
+                target_update_freq=500,
                 action_space=env.action_space,
                 is_double=True,
             ).to("cuda" if torch.cuda.is_available() else "cpu")
@@ -82,16 +110,56 @@ def _get_agents(
                 model=net_fixed,
                 optim=optim_fixed,
                 # discount_factor=0.98,
-                estimation_step=15,
-                target_update_freq=1000000000,
+                estimation_step=5,
+                target_update_freq=500,
                 action_space=env.action_space,
                 is_double=True,
             ).to("cuda" if torch.cuda.is_available() else "cpu")
+        if icm and not is_distrib:
+            feature_dim = net.output_dim
+            feature_net = make_net(features_only=True)
+            feature_net_fixed = make_net(features_only=True).train(False)
+            icm_net = IntrinsicCuriosityModule(
+                feature_net=net,
+                feature_dim=feature_dim,
+                action_dim=action_dim,
+                hidden_sizes=hidden_sizes,
+                device="cuda" if torch.cuda.is_available() else "cpu",
+            )
+            icm_net_fixed = IntrinsicCuriosityModule(
+                feature_net=net_fixed,
+                feature_dim=feature_dim,
+                action_dim=action_dim,
+                hidden_sizes=hidden_sizes,
+                device="cuda" if torch.cuda.is_available() else "cpu",
+            )
+            icm_lr_scale = 1.0
+            icm_reward_scale = 0.01
+            icm_forward_loss_weight = 0.2
+            agent_learn = ICMPolicy(
+                policy=agent_learn,
+                model=icm_net,
+                optim=icm_optim,
+                action_space=env.action_space,
+                lr_scale=icm_lr_scale,
+                reward_scale=icm_reward_scale,
+                forward_loss_weight=icm_forward_loss_weight,
+            ).to("cuda" if torch.cuda.is_available() else "cpu")
+            agent_fixed = ICMPolicy(
+                policy=agent_fixed,
+                model=icm_net_fixed,
+                optim=optim_fixed,
+                action_space=env.action_space,
+                lr_scale=icm_lr_scale,
+                reward_scale=icm_reward_scale,
+                forward_loss_weight=icm_forward_loss_weight,
+            ).to("cuda" if torch.cuda.is_available() else "cpu")
+
         if is_distrib:
-            agent_learn = RainbowPolicy(model=net, optim=optim, action_space=env.action_space, estimation_step=10,
-                                        target_update_freq=5000).to("cuda" if torch.cuda.is_available() else "cpu")
+            agent_learn = RainbowPolicy(model=net, optim=optim, action_space=env.action_space, estimation_step=4,
+                                        target_update_freq=1000).to("cuda" if torch.cuda.is_available() else "cpu")
             agent_fixed = RainbowPolicy(model=net_fixed, optim=optim_fixed, action_space=env.action_space,
-                                        estimation_step=10, target_update_freq=5000).to(
+                                        estimation_step=4, target_update_freq=1000).to(
                 "cuda" if torch.cuda.is_available() else "cpu")
 
     from tianshou.data import Batch
@@ -130,6 +198,10 @@ def watch_self_play():
     policy.policies[ag[0]].load_state_dict(policy.policies[ag[1]].state_dict().copy())
     policy.policies[ag[1]].set_eps(0.01)
     policy.policies[ag[0]].set_eps(0.01)
+    if os.path.islink("log/ttt/dqn/policy.pth"):
+        link_target = os.readlink("log/ttt/dqn/policy.pth")
+    else:
+        link_target = None
     while True:
         obs, _ = env.reset()
         p1_obs = obs["0"]["observation"]
@@ -149,18 +221,30 @@ def watch_self_play():
             act1 = policy.policies[ag[0]].compute_action(p1_obs)
             act2 = policy.policies[ag[1]].compute_action(p2_obs)
             if done["0"]:
+                if link_target is not None:
+                    if link_target != os.readlink("log/ttt/dqn/policy.pth"):
+                        return
                 break
 
 
 if __name__ == "__main__":
+    # get commandline args
+    parser = argparse.ArgumentParser()
+    parser.add_argument("--watch", action="store_true", help="Watch the trained agent play")
+    args = parser.parse_args()
+    if args.watch:
+        while True:
+            watch_self_play()
+
+
     # ======== Step 1: Environment setup =========
     def with_render(env):
         env.always_render = True
         return env
 
 
-    train_envs = DummyVectorEnv([_get_env for _ in range(10)])
-    test_envs = DummyVectorEnv([_get_env for _ in range(1)])
+    train_envs = SubprocVectorEnv([_get_env for _ in range(10)])
+    test_envs = SubprocVectorEnv([_get_env for _ in range(10)])
 
     # seed
     seed = 1
@@ -170,34 +254,38 @@ if __name__ == "__main__":
     test_envs.seed(seed)
 
     # ======== Step 2: Agent setup =========
-    policy, optim, agents = _get_agents()
+    policy, optim, agents = _get_agents(league=False)
 
     # ======== Step 3: Collector setup =========
     train_collector = Collector(
         policy,
         train_envs,
         # HERVectorReplayBuffer(20_000, len(train_envs), compute_reward_fn=lambda x: x[1], horizon=1000),
-        VectorReplayBuffer(20_000, len(train_envs)),
+        VectorReplayBuffer(100_000, len(train_envs)),
+        # PrioritizedVectorReplayBuffer(80_000, len(train_envs), alpha=0.6, beta=0.4, weight_norm=True, ignore_obs_next=True),
         exploration_noise=True,
     )
     test_collector = Collector(policy, test_envs, exploration_noise=True)
     # policy.set_eps(1)
     train_collector.collect(n_step=64 * 10)  # batch size * training_num
+
+
     def update_league_and_forward(self, fun, *args, **kwargs):
         for pol in self.policy.policies.values():
             if hasattr(pol, "update_idx"):
                 pol.update_idx()
         fun(*args, **kwargs)
-    test_collector.reset_env__ = test_collector.reset_env
-    test_collector._reset_env_with_ids__ = test_collector._reset_env_with_ids
-    test_collector.reset_env = lambda *args, **kwargs: update_league_and_forward(test_collector, test_collector.reset_env__, *args, **kwargs)
-    test_collector._reset_env_with_ids = lambda *args, **kwargs: update_league_and_forward(test_collector, test_collector._reset_env_with_ids__, *args, **kwargs)
+
+
+    # test_collector.reset_env__ = test_collector.reset_env
+    # test_collector._reset_env_with_ids__ = test_collector._reset_env_with_ids
+    # test_collector.reset_env = lambda *args, **kwargs: update_league_and_forward(test_collector, test_collector.reset_env__, *args, **kwargs)
+    # test_collector._reset_env_with_ids = lambda *args, **kwargs: update_league_and_forward(test_collector, test_collector._reset_env_with_ids__, *args, **kwargs)
 
     # train_collector.reset_env__ = train_collector.reset_env
     # train_collector._reset_env_with_ids__ = train_collector._reset_env_with_ids
     # train_collector.reset_env = lambda *args, **kwargs: update_league_and_forward(train_collector, train_collector.reset_env__, *args, **kwargs)
     # train_collector._reset_env_with_ids = lambda *args, **kwargs: update_league_and_forward(test_collector, train_collector._reset_env_with_ids__, *args, **kwargs)
-
 
     def override_kwargs_and_forward(fun, *args, **kwargs):
         kwargs["render"] = 0.00000001
@@ -208,40 +296,47 @@ if __name__ == "__main__":
     # test_collector.collect___r = lambda *args, **kwargs: override_kwargs_and_forward(test_collector.collect___b, *args, **kwargs)
     # test_collector.collect = test_collector.collect___r
 
-    thresh_win_rate = 0.7
-    thresh_in_a_row = 10
+    thresh_win_rate = 0.6
+    thresh_in_a_row = 3
     in_a_row = 0
     wins = 0
     losses = 0
     draws = 0
 
     # ======== Step 4: Callback functions setup =========
-    import multiprocessing
-
-    multiprocessing.set_start_method("spawn")
-    p = multiprocessing.Process(target=watch_self_play)
+    if watch:
+        # multiprocessing.set_start_method("spawn")
+        p = multiprocessing.Process(target=watch_self_play)
 
 
     def save_best_fn(policy):
         global p
-        model_save_path = os.path.join("log", "ttt", "dqn", "policy.pth")
+        # get datetime
+        timestr = datetime.datetime.now().strftime("%Y%m%d-%H%M%S")
+        model_save_path = os.path.join(os.getcwd(), "log", "ttt", "dqn", f"policy-{timestr}.pth")
         os.makedirs(os.path.join("log", "ttt", "dqn"), exist_ok=True)
         torch.save(policy.policies[agents[1]].state_dict(), model_save_path)
+        # make symlink
+        os.remove(os.path.join(os.getcwd(), "log", "ttt", "dqn", "policy.pth"))
+        os.symlink(model_save_path, os.path.join(os.getcwd(), "log", "ttt", "dqn", "policy.pth"))
         # create new process to watch the trained agent play
-        while p.is_alive():
-            p.terminate()
-        p = multiprocessing.Process(target=watch_self_play)
-        p.start()
+        if watch:
+            while p.is_alive():
+                p.terminate()
+            p = multiprocessing.Process(target=watch_self_play)
+            p.start()
 
 
     def stop_fn(mean_rewards):
         global wins, losses, draws, in_a_row
-        winrate = wins / (wins + losses + draws)
+        winlossrate = wins / (wins + losses + 1e-6)
+        winrate = wins / (wins + draws + losses)
+        print(f"win/loss rate: {winlossrate}, win rate: {winrate}, wins: {wins}, losses: {losses}, draws: {draws}")
         wins = 0
         losses = 0
         draws = 0
-        print(f"winrate: {winrate}")
-        if winrate > thresh_win_rate:
+
+        if winlossrate > thresh_win_rate:
             in_a_row += 1
         else:
             in_a_row = 0
@@ -250,9 +345,14 @@ if __name__ == "__main__":
 
     def train_fn(epoch, env_step):
         global wins, draws, losses
-        policy.policies[agents[1]].set_eps(0.01)
-        policy.policies[agents[0]].update_idx()
-        policy.policies[agents[0]].set_eps(0.01)
+        eps = 0.2
+        if env_step > 100_000:
+            eps = 0.1
+            eps -= 0.002 * (env_step - 100_000)
+            eps = max(0.01, eps)
+        policy.policies[agents[1]].set_eps(eps)
+        # policy.policies[agents[0]].update_idx()
+        policy.policies[agents[0]].set_eps(eps)
 
 
     def test_fn(epoch, env_step):
@@ -275,16 +375,16 @@ if __name__ == "__main__":
             policy=policy,
             train_collector=train_collector,
             test_collector=test_collector,
-            max_epoch=4000,
-            step_per_epoch=10_000,
-            step_per_collect=100,
-            episode_per_test=8,
-            batch_size=32,
+            max_epoch=5000,
+            step_per_epoch=20_000,
+            step_per_collect=10,
+            episode_per_test=20,
+            batch_size=16,
             train_fn=train_fn,
             test_fn=test_fn,
             stop_fn=stop_fn,
             save_best_fn=save_best_fn,
-            update_per_step=.3,
+            update_per_step=.08,
             test_in_train=False,
             reward_metric=reward_metric,
         ).run()
@@ -292,12 +392,12 @@ if __name__ == "__main__":
         losses = 0
         draws = 0
         in_a_row = 0
-        agent_copy = copy.copy(policy.policies[agents[1]])
-        agent_copy.load_state_dict(policy.policies[agents[1]].state_dict().copy())
-        policy.policies[agents[0]].push_agent(agent_copy, rotate=True)
+        # agent_copy = copy.copy(policy.policies[agents[1]])
+        # agent_copy.load_state_dict(policy.policies[agents[1]].state_dict().copy())
+        # policy.policies[agents[0]].push_agent(agent_copy, rotate=True)
 
         print(f"Generation {i + 1}/{gen_count} complete")
-        # policy.policies[agents[0]].load_state_dict(policy.policies[agents[1]].state_dict().copy())
+        policy.policies[agents[0]].load_state_dict(policy.policies[agents[1]].state_dict().copy())
 
     # return result, policy.policies[agents[1]]
     print(f"\n==========Result==========\n{result}")
